reorganize where card components will be housed
